requirements.txt
pdf2image
Pillow
matplotlib
numpy
jiwer
python-Levenshtein
rapidfuzz
tqdm
easyocr
pytesseract
# paddlepaddle and paddleocr install instructions are in README (Paddle has macOS-specific wheels)
google-genai
tqdm
python-dotenv

pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
pydantic-core==2.20.1
typing_extensions
annotated_types
google-auth
anyio
sniffio
certifi
requests
urllib3
idna
cryptography
pyasn1
pyasn1_modules
rsa
cachetools
httpx
tenacity
websockets
httpcore
h11
packaging


sympy
mpmath
opencv-python
scipy
scikit-image
lazy_loader
tifffile
python-bidi
pyyaml
Levenshtein
jiwer==2.6.0

huggingface_hub
filelock
regex
safetensors
colorlog
fsspec==2025.9.0
jinja2==3.1.6
markupsafe==3.0.2
networkx==3.5
setuptools==80.9.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
accelerate==1.10.1
charset-normalizer==3.4.3
hf-xet==1.1.9
psutil==7.0.0
tokenizers==0.22.0
transformers==4.56.1